+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable and major changes to this project will be documented in this file.
 
 ### Added
 - ROKS - Restricted Open Shell Kohn Sham DFT `q_block\compute\solvers\electronic_density\kohn_sham\restricted_open_shell_kohn_sham.py` with tests.
+- Sphinx documentation  intialization
 
 ### Changed
 - ...